Fixed URL handler that would truncate params with "." in them
Default Spring pattern matching for parameters assumes that "."
should be truncated from parameters at the end of URLs.
The better fix would be to handle it in the application config, but
the suggested DefaultAnnotationHandlerMapping is deprecated and the
RequestMappingHandlerMapping that replaces it does not work the
same way. Instead, I modified the pattern matching at the end of
each API URL to include zero or more periods.